requirements.txt
tensorflow
numpy
flask
keras
flask-sqlalchemy
pillow